pyproject.toml
[project]
name = "soup-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "dj-database-url>=3.0.1",
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "django-ninja>=1.4.3",
    "django-ninja-extra>=0.30.1",
    "dotenv>=0.9.9",
    "gunicorn>=23.0.0",
    "ninja-schema>=0.14.3",
    "psycopg[binary]>=3.2",
    "pydantic==2.11.9",
    "pyjwt>=2.10.1",
    "requests>=2.32.5",
    "social-auth-app-django>=5.6.0",
    "uuid6>=2025.0.1",
]

[dependency-groups]
dev = [
    "coverage>=7.11.0",
    "faker>=37.11.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = "tests.py test_*.py *_tests.py"
